Create README.md
This repository belongs to my review paper "Kinetics of heat-induced changes in foods: a workflow proposal" published in the Journal of Food Engineering, 2021 (https://doi.org/10.1016/j.jfoodeng.2021.110634). It uses bayesian regression with the R package brms and frequentist linear and nonlinear regression. The directory is structured of a data folder, brms fit files folder and the R script.